requirements.txt
-i https://pypi.org/simple
certifi==2020.6.20
chardet==3.0.4
deprecated==1.2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
idna==2.10; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3'
pygithub==1.51
pyjwt==1.7.1
requests==2.24.0
urllib3==1.25.9; python_version >= '2.7' and python_version not in '3.0, 3.1, 3.2, 3.3, 3.4' and python_version < '4'
wrapt==1.12.1
